+++ crates/openshell-cli/src/lib.rs
@@ -9,6 +9,7 @@
 pub(crate) static TEST_ENV_LOCK: std::sync::Mutex<()> = std::sync::Mutex::new(());
 
 pub mod auth;
+pub mod commands;
 pub mod completers;
 pub mod edge_tunnel;
 pub mod oidc_auth;
